Added Stack project
The Stack class demonstrates using Exceptions.
It also demonstrates generics again – the stack class is a generic class.
We also showed that if a generic class has a static member – a different copy of the static member will be created for each instantiation of the generic class.
In other words, Stack<int> and Stack<double> will each have their own static member.